Readable: repair label font, ribbon label, tag label
Because the label in Semantic-UI has no border, adding a border to it
causes tricky problems in magic labels such as tag label and ribbon label.

Tag label: the small triangle on the left is actually a transformed
square, after adding border to the main body of tag, totally 2 pixels,
the diagonal of the square is shorter than the height of the tag's height,
so I increase the width and height of the left small square slightly(1.56em -> 1.62em),
and move it to left by 1%, thus it may looks better, but in fact, it's
different from the originl design.

Ribbon label: the ribbon label's shadow is the top right half of a small
square, it's almost impossible to add a border to the diagonal(maybe
there's a way to do that, I do not investegate yet at present), so I move the
square to left by 1px, and set the background color slightly
gray(darken(white, 5%)), so it may display in harmony with the ribbon
label body.