Migrate build script from bash to node.js.
As in #13, we use bash to build css before, however, bash script is not
portable. so we have to migrate the build script from bash to node and
make sure it builds on windows platform.

The goal of the new build script is that we should finish all the build
steps via `npm`, that means, the basic steps to build should be:

- `npm install`, and then
- `npm build`, and then you get all themes

To meet this goal, we have to add semantic-ui as a npm dependency in
`package.json`, and when you `npm install semantic-ui` or `yarn add
semantic-ui`, the install progress will import many some source code of
semantic-ui into your working directory, that's why this commit is so
huge that contains lots of source code from semantic-ui.